refactor: Update transfer config hook path
